replace modal button from index to 'ModalMain' component
